+++ dust-compiler.js
@@ -3,16 +3,27 @@
 
 'use strict';
 
-var source = 'change-me/',      // must end in slash
-    destination = 'change-me/', // must end in slash
-    fs = require('fs'),
+var fs = require('fs'),
     log,
     path = require('path'),
-    mkdirp = require('mkdirp'),
+    argv = require('optimist')
+        .usage('Usage: dust-compiler -s|--source source_path -d|--destination destination_path [--bootstrap] [--nonotify] [--includepath]')
+        .demand(['source', 'destination'])
+        .alias('source', 's')
+        .alias('destination', 'd')
+        .describe('source', 'The source .dust templates to watch.')
+        .describe('destination', 'The destination destination to write compiles .js files out to.')
+        .describe('bootstrap', 'Does not replicate.  Displays the curl command that would be used.')
+        .describe('nonotify', 'Do not notify.')
+        .describe('includepath', 'Include the path in the name of the compiled dust template.')
+        .argv,
     dust = require('dustjs-linkedin'),
     watch = require('watch'),
+    colors = require('colors'),
+    mkdirp = require('mkdirp'),
     wrench = require('wrench'),
-    bootstrap = false;
+    source = argv.s.match('/$') ? path.normalize(argv.s) : path.normalize(argv.s + '/'),      // must end in slash
+    destination = argv.d.match('/$') ? path.normalize(argv.d) : path.normalize(argv.d + '/'); // must end in slash
 
 
 log = (function () {
@@ -22,7 +33,7 @@ log = (function () {
     case 'darwin':
         notifier = require('terminal-notifier');
         return function (message) {
-            notifier(message, {
+            notifier(message.stripColors, {
                 title: 'Dust Compiler',
                 activate: 'com.apple.Terminal'
             });
@@ -32,7 +43,7 @@ log = (function () {
     case 'linux':
         notifier = require('notify-send');
         return function (message) {
-            notifier.notify('Dust Compiler', message);
+            notifier.notify('Dust Compiler', message.stripColors);
             console.log(message);
         };
 
@@ -42,35 +53,36 @@ log = (function () {
 }());
 
 
-
 function compile(src, curr, prev) {
-    var filename,
-        filepath,
+    var data,
         error = false,
-        data;
+        filename,
+        filepath,
+        compileFilename;
 
     src = path.normalize(src);
 
     if (path.extname(src) === '.dust') {
         fs.stat(src, function (err, stat) {
             if (err) {
-                log('fs.stat error');
+                log('[ERROR]'.red + ' fs.stat');
                 throw err;
             }
 
             if (!stat.isDirectory()) {
 
                 filename = src.substring(source.length);
+                compileFilename = argv.includepath ? destination + filename : filename;
                 filepath = destination + filename.substring(0, filename.length - 5) + '.js';
 
                 fs.readFile(src, function (err, data) {
                     if (err) {
-                        log('fs.readFile error');
+                        log('[ERROR]'.red + ' fs.readFile');
                         throw err;
                     }
 
                     try {
-                        data = dust.compile(data.toString(), filename);
+                        data = dust.compile(data.toString(), compileFilename);
                     } catch (e) {
                         error = true;
                         log(e);
@@ -79,16 +91,16 @@ function compile(src, curr, prev) {
                     if (!error) {
                         mkdirp.mkdirp(path.dirname(filepath), function (err) {
                             if (err) {
-                                log('mkdirp.mkdirp error: ' + err);
+                                log('[ERROR]'.red + ' mkdirp.mkdirp: ' + err);
                                 throw err;
                             }
 
                             fs.writeFile(filepath, data, function (err) {
                                 if (err) {
-                                    log('fs.writeFile error: ' + err);
+                                    log('[ERROR]'.red + ' fs.writeFile: ' + err);
                                     throw err;
                                 }
-                                log('Saved ' + filepath);
+                                log('Compiled '.green + filepath + ' as '.green + compileFilename);
                             });
                         });
                     }
@@ -99,29 +111,12 @@ function compile(src, curr, prev) {
 }
 
 
-process.argv.forEach(function (arg, idx) {
-    if (idx < 2) { return; }
-
-    switch (arg) {
-    case '--bootstrap':
-        bootstrap = true;
-        break;
-
-    case '--no-notify':
-        log = console.log;
-        break;
-
-    default:
-        log('Ignoring unrecognized option: ' + arg);
-    }
-});
+if (argv.nonotify) {
+    log = console.log;
+}
 
 
-source = path.normalize(source);
-destination = path.normalize(destination);
-
-
-if (bootstrap) {
+if (argv.bootstrap) {
     wrench.readdirRecursive(source, function (error, fileList) {
         if (fileList) {
             fileList.forEach(function (filename) {
@@ -129,10 +124,26 @@ if (bootstrap) {
             });
         }
     });
-} else {
-    watch.createMonitor(source, function (monitor) {
-        log('Watching ' + source);
-        monitor.on('created', compile);
-        monitor.on('changed', compile);
-    });
 }
+
+
+if (!fs.existsSync(source)) {
+    console.log('[ERROR]'.red + ' ' + source + ' does not exist');
+    process.exit();
+}
+
+
+if (!fs.existsSync(destination)) {
+    console.log('[ERROR]'.red + ' ' + destination + ' does not exist');
+    process.exit();
+}
+
+
+watch.createMonitor(source, function (monitor) {
+    log('Watching '.green + source);
+    log('Will write to '.yellow + destination.yellow);
+    log('Type ^C to cancel'.yellow);
+
+    monitor.on('created', compile);
+    monitor.on('changed', compile);
+});